# docs/api/openapi-draft.yaml
openapi: 3.1.0
info:
  title: SpendGuard API
  version: 0.1.0-draft
  description: Day 1 API contract. Implementation starts after review.
servers:
  - url: http://localhost:8080
...
paths:
  /api/v1/spend-events:
    post:
      tags: [Spend Events]
      summary: Submit a financial notification text
      operationId: submitSpendEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitSpendEventRequest'
      responses:
        '202':
          description: Event accepted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendEventAcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Duplicate event
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/spend-events/{eventId}:
    get:
      tags: [Spend Events]
      summary: Get processing status and result
      operationId: getSpendEvent
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Current event state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendEventDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    SubmitSpendEventRequest:
      type: object
      required: [source, message]
      properties:
        source:
          type: string
          enum: [MANUAL_TEXT, SIMULATOR]
        externalEventId:
          type: [string, 'null']
          maxLength: 200
        message:
          type: string
          minLength: 1
          maxLength: 2000
        occurredAt:
          type: [string, 'null']
          format: date-time
    SpendEventAcceptedResponse:
      type: object
      required: [eventId, status, receivedAt]
      properties:
        eventId:
          type: string
          format: uuid
        status:
          type: string
          const: RECEIVED
        receivedAt:
          type: string
          format: date-time
    SpendEventDetail:
      type: object
      required: [eventId, status, receivedAt]
      properties:
        eventId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RECEIVED, PUBLISHED, ANALYZING, COMPLETED, NEEDS_REVIEW, RETRY_WAIT, DEAD_LETTERED]
        receivedAt:
          type: string
          format: date-time
        transaction:
          oneOf:
            - $ref: '#/components/schemas/Transaction'
            - type: 'null'
        risk:
          oneOf:
            - $ref: '#/components/schemas/RiskAssessment'
            - type: 'null'
    Transaction:
      type: object
      required: [id, amount, currency, transactionType, category, confidence, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0
        currency:
          type: string
          default: KRW
        transactionType:
          type: string
          enum: [PAYMENT, CANCEL, REFUND]
        merchant:
          type: [string, 'null']
        normalizedMerchant:
          type: [string, 'null']
        category:
          $ref: '#/components/schemas/Category'
        confidence:
          type: number
          minimum: 0
          maximum: 1
        occurredAt:
          type: string
          format: date-time
    Category:
      type: string
      enum: [FOOD, TRANSPORTATION, SHOPPING, SUBSCRIPTION, HOUSING, HEALTH, EDUCATION, LEISURE, TRANSFER, OTHER]
...
    RiskAssessment:
      type: object
      required: [score, level, policyVersion, signals]
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        level:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        policyVersion:
          type: string
        signals:
          type: array
          items:
            type: object
            required: [type, message]
            properties:
              type:
                type: string
              message:
                type: string
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string